Avoid requiring the REPL to be loaded to show error hints for undefined variable
